requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.7.4.post0
anyio==3.3.4; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
fastapi==0.70.0
feedparser==6.0.8
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
h11==0.12.0; python_version >= '3.6'
httptools==0.2.0
idna==3.3; python_version >= '3'
iniconfig==1.1.1
lunr==0.6.1
mmh3==3.0.0
multidict==5.2.0; python_version >= '3.6'
oauthlib==3.1.1; python_version >= '3.6'
packaging==21.0; python_version >= '3.6'
pluggy==1.0.0; python_version >= '3.6'
pottery==1.4.6
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.8.2; python_full_version >= '3.6.1'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-asyncio==0.16.0
pytest==6.2.5
python-bugzilla==3.1.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-twitter==3.5
pyyaml==6.0
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sgmllib3k==1.0.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
soupsieve==2.2.1; python_version >= '3.6'
starlette==0.16.0; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.2
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn[standard]==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
yarl==1.7.0; python_version >= '3.6'
